scst_vdisk: Remove a superfluous cast
git-svn-id: http://svn.code.sf.net/p/scst/svn/trunk@7044 d57e44dd-8a1f-0410-8b47-8ef2f437770f

+++ scst/src/dev_handlers/scst_vdisk.c
@@ -2048,7 +2048,7 @@ static int vdisk_format_dif(struct scst_cmd *cmd, uint64_t start_lba,
 				 iv_count, &loff, 0);
 		if (err < 0) {
 			PRINT_ERROR("Formatting DIF write() returned %lld from "
-				"%zd", (long long)err, full_len);
+				"%zd", err, full_len);
 			if (err == -EAGAIN)
 				scst_set_busy(cmd);
 			else
